requirements.txt
ipykernel==6.25.2
joblib==1.3.2
keras==2.14.0
numpy==1.26.1
pandas==2.1.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.3
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
urllib3==2.0.7
keras-tuner==1.4.5